Remove Tiedonsiirto-related endpoints from backend
